requirements.txt
Flask==3.0.0
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
mediapipe==0.10.9
onnxruntime==1.17.0